pyproject.toml
[project]
name = "ADM"
version = "0.1.0"
description = "An Web-Site to administrate your life"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "dj-database-url>=3.0.1",
    "django==5.2.3",
    "django-colorfield>=0.14.0",
    "django-mathfilters>=1.0.0",
    "google>=3.0.0",
    "google-genai>=1.29.0",
    "gunicorn>=23.0.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "sqlparse==0.5.3",
    "whitenoise>=6.9.0",
]
